Remove loader error dialog * Always show ROMs * Address CRs

+++ app/src/main/cpp/skyline/vfs/ctr_encrypted_backing.cpp
@@ -8,7 +8,7 @@ namespace skyline::vfs {
 
     CtrEncryptedBacking::CtrEncryptedBacking(crypto::KeyStore::Key128 &ctr, crypto::KeyStore::Key128 &key, const std::shared_ptr<Backing> &backing, size_t baseOffset) : Backing({true, false, false}), ctr(ctr), cipher(key, MBEDTLS_CIPHER_AES_128_CTR), backing(backing), baseOffset(baseOffset) {}
 
-    void CtrEncryptedBacking::UpdateCtr(size_t offset) {
+    void CtrEncryptedBacking::UpdateCtr(u64 offset) {
         offset >>= 4;
         size_t le{__builtin_bswap64(offset)};
         std::memcpy(ctr.data() + 8, &le, 8);
